requirements.txt
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.1.5
Pillow==6.0.0
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
